Fix issue with multiple text file inclusion
If multiple `@[]()` lines were used in a single lab file (to
include multiple text files), there was an error in the regex
matching that would group together the two lines in a single
match and mangle everything.

Updated the regex so that the text would not match any newlines
(and thus everything would be on the same line).