Fix issue with two consecutive text file inclusions
If you had two text inclusions, the first text inclusion would
"eat" the newline required for the regex of the second text
inclusion. This fixes that problem by requiring a single newline
at the end. Thus, two sequential text inclusions work as long
as there is a space in between them.

I honestly don't remember why the text inclusion needs to consume
the newline at the end. Now it ensures that the text inclusion
is alone on the line, but it was not that way before. Nevertheless,
I am hesitent to remove that as I don't want to cause problems
with existing notes.